Temporarily Reinstate Akka DNS as normal dependency
This is done to mitigate the fact that tooling has been released with the following problem.

If a non-Lagom project enables a service locator, the shaded Reactive Lib will fail with the following error.

```
[error] missing or invalid dependency detected while loading class file 'ServiceLocatorLike.class'.
[error] Could not access term ru in package <root>,
[error] because it (or its dependencies) are missing. Check your build definition for
[error] missing or conflicting dependencies. (Re-run with `-Ylog-classpath` to see the problematic classpath.)
[error] A full rebuild may help if 'ServiceLocatorLike.class' was compiled against an incompatible version of <root>.
```

This revert is done so the tooling that's released will at least work with Akka DNS as dependency while we are working to ensure the issue above does not occur with Akka DNS shading.